feat: updated proto definitons and simplified wasm code
 -proto: making each field a u32 does not mean that it all takes 4 bytes but rather gets compressed to use only one if it does not exceed the values of a u8

+++ memoria/server/src/game.rs
@@ -12,9 +12,9 @@ use tonic::{Request, Response, Status};
 
 #[derive(Debug)]
 pub struct Game {
-    pub width: u16,
-    pub height: u16,
-    pub player_cap: u32,
+    pub width: u8,
+    pub height: u8,
+    pub player_cap: u8,
     pub connected: Vec<Sender<Result<GameStateReply, Status>>>,
 }
